app/requirements.txt
flask==3.0.3
python-jose[cryptography]==3.3.0
requests==2.31.0
gunicorn==22.0.0
